+++ oshi-core/src/main/java/oshi/software/os/mac/MacOperatingSystem.java
@@ -210,6 +210,40 @@ public class MacOperatingSystem extends AbstractOperatingSystem {
         return proc;
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public OSProcess[] getChildProcesses(int parentPid, int limit, ProcessSort sort) {
+        List<OSProcess> procs = new ArrayList<>();
+        int[] pids = new int[this.maxProc];
+        int numberOfProcesses = SystemB.INSTANCE.proc_listpids(SystemB.PROC_ALL_PIDS, 0, pids,
+                pids.length * SystemB.INT_SIZE) / SystemB.INT_SIZE;
+        for (int i = 0; i < numberOfProcesses; i++) {
+            // Handle off-by-one bug in proc_listpids where the size returned
+            // is: SystemB.INT_SIZE * (pids + 1)
+            if (pids[i] == 0) {
+                continue;
+            }
+            if (parentPid == getParentProcessPid(pids[i])) {
+                OSProcess proc = getProcess(pids[i]);
+                if (proc != null) {
+                    procs.add(proc);
+                }
+            }
+        }
+        List<OSProcess> sorted = processSort(procs, limit, sort);
+        return sorted.toArray(new OSProcess[sorted.size()]);
+    }
+
+    private int getParentProcessPid(int pid) {
+        ProcTaskAllInfo taskAllInfo = new ProcTaskAllInfo();
+        if (0 > SystemB.INSTANCE.proc_pidinfo(pid, SystemB.PROC_PIDTASKALLINFO, 0, taskAllInfo, taskAllInfo.size())) {
+            return 0;
+        }
+        return taskAllInfo.pbsd.pbi_ppid;
+    }
+
     private String getCommandLine(int pid) {
         // Get command line via sysctl
         int[] mib = new int[3];